+++ src/Protsyk.RayTracer.Challenge.Core/Scene/Lights/SpotLight.cs
@@ -25,7 +25,7 @@ namespace Protsyk.RayTracer.Challenge.Core.Scene.Lights
             return Tuple4.Subtract(location, from).Length();
         }
 
-        public double GetIntensity(Tuple4 dir, int shine, Tuple4 pointOnSurface, Tuple4 surfaceNormal)
+        public double GetIntensity(Tuple4 dir, double shine, Tuple4 pointOnSurface, Tuple4 surfaceNormal)
         {
             var lightDirection = Tuple4.Normalize(Tuple4.Subtract(location, pointOnSurface));
             var result = 0.0;
